Quiz-App-with-Sounds-AndroidStudio
I created two music quizzes with 11 questions, one in English and one in Turkish, in Android Studio. In the quizzes, the user should listen to the song by pressing the play button, then answer the question and press the next button to go to the next question. The user will be able to see his score and the best score at the end of the quiz.